# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My PDF Converter API",
        "description": "Upload any document and get a PDF in response",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/password/confirm": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify the authenticated user’s password",
                "description": "Confirm the user’s current password for sensitive actions.",
                "operationId": "7acc552047719feda4ebbe3d9f7d7e9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "currentPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password confirmed"
                    },
                    "422": {
                        "description": "Invalid password"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/email/verification-notification": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send email verification link",
                "description": "Send a new email verification notification to the authenticated user.",
                "operationId": "21ff3f9f9b9e1407909089f864312e1c",
                "responses": {
                    "200": {
                        "description": "Verification link sent"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/password": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change the authenticated user’s password",
                "description": "Update the current user’s password.",
                "operationId": "5609218cd711f2685cb3a76db17ee277",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldSecret123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password successfully updated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mark the authenticated user’s email as verified",
                "description": "Verify a user’s email address via signed link.",
                "operationId": "28405e7cec5bd7034f5ee5b3261b3cad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "SHA1 of the user’s email address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to dashboard with `?verified=1`"
                    },
                    "403": {
                        "description": "Invalid or expired verification link"
                    }
                }
            }
        },
        "/api/login-history": {
            "delete": {
                "tags": [
                    "LoginHistory"
                ],
                "summary": "Truncate the login history (admin only)",
                "description": "Delete *all* login history records.",
                "operationId": "9635b590cc8365e3595b8a95b576294a",
                "responses": {
                    "204": {
                        "description": "All login records deleted"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login-history/export": {
            "get": {
                "tags": [
                    "LoginHistory"
                ],
                "summary": "Download login history CSV (admin only)",
                "description": "Export the login history as a CSV download.",
                "operationId": "cac354680d9a4cfb7f0a28170a72e572",
                "responses": {
                    "200": {
                        "description": "CSV file stream",
                        "content": {
                            "text/csv": {}
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}